Added data tag to all generated filenames
The data tag is in the format of `date +%Y%m%d`, to be consistent with
filenames in current aarch-system-2014-10.tar.xz.

Also renamed one of the DTBs for the VExpress_EMM platform to include
the number of CPUs it is desined for.

Modified README.md accordingly with the new filenames, and added some
new information about the differences with aarch-system-2014-10.tar.xz.

Added copyright to run.sh